+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-01T09:48:20.380Z",
+    "2026-05-02T11:03:52.191Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {