+++ packages/orq-rc/RELEASES.md
@@ -8799,3 +8799,13 @@ Based on:
 - [typescript v4.9.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.17 - packages/orq-rc
+
+## 2026-05-02 11:07:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.18 - packages/orq-rc